Extended usage of `service/internal/crud`
Using it in both the securities and portfolio service.

Fixes #43
