+++ service/securities/service.go
@@ -18,7 +18,6 @@
 package securities
 
 import (
-	"context"
 	"time"
 
 	moneygopher "github.com/oxisto/money-gopher"
@@ -26,17 +25,11 @@ import (
 	"github.com/oxisto/money-gopher/gen/portfoliov1connect"
 	"github.com/oxisto/money-gopher/persistence"
 
-	"github.com/bufbuild/connect-go"
-	"golang.org/x/exp/slices"
 	"golang.org/x/text/currency"
-	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 type service struct {
-	// TODO(oxisto): convert this to sqlite
-	sec map[string]*portfoliov1.Security
-
 	securities       persistence.StorageOperations[*portfoliov1.Security]
 	listedSecurities persistence.StorageOperations[*portfoliov1.ListedSecurity]
 
@@ -82,76 +75,4 @@ func NewService(db *persistence.DB) portfoliov1connect.SecuritiesServiceHandler 
 		securities:       securities,
 		listedSecurities: listedSecurities,
 	}
-}
-
-func (svc *service) CreateSecurity(ctx context.Context, req *connect.Request[portfoliov1.CreateSecurityRequest]) (res *connect.Response[portfoliov1.Security], err error) {
-	svc.sec[req.Msg.Security.Name] = req.Msg.Security
-	return
-}
-
-func (svc *service) GetSecurity(ctx context.Context, req *connect.Request[portfoliov1.GetSecurityRequest]) (res *connect.Response[portfoliov1.Security], err error) {
-	res = connect.NewResponse(&portfoliov1.Security{})
-	res.Msg, err = svc.fetchSecurity(req.Msg.Name)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-
-	return
-}
-
-func (svc *service) ListSecurities(ctx context.Context, req *connect.Request[portfoliov1.ListSecuritiesRequest]) (res *connect.Response[portfoliov1.ListSecuritiesResponse], err error) {
-	res = connect.NewResponse(&portfoliov1.ListSecuritiesResponse{})
-	res.Msg.Securities, err = svc.securities.List()
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-
-	for _, sec := range res.Msg.Securities {
-		sec.ListedOn, err = svc.listedSecurities.List(sec.Name)
-		if err != nil {
-			return nil, connect.NewError(connect.CodeInternal, err)
-		}
-	}
-
-	return
-}
-
-func (svc *service) UpdateSecurity(ctx context.Context, req *connect.Request[portfoliov1.UpdateSecurityRequest]) (res *connect.Response[portfoliov1.Security], err error) {
-	res = connect.NewResponse(&portfoliov1.Security{})
-	res.Msg, err = svc.securities.Update(req.Msg.Security.Name, req.Msg.Security, req.Msg.UpdateMask.Paths)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-
-	if slices.Contains(req.Msg.UpdateMask.Paths, "listed_on") {
-		for _, ls := range req.Msg.Security.ListedOn {
-			svc.listedSecurities.Replace(ls)
-		}
-	}
-
-	return
-}
-
-func (svc *service) DeleteSecurityRequest(ctx context.Context, req *connect.Request[portfoliov1.DeleteSecurityRequest]) (res *connect.Response[emptypb.Empty], err error) {
-	res = connect.NewResponse(&emptypb.Empty{})
-	err = svc.securities.Delete(req.Msg.Name)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-
-	return
-}
-
-func (svc *service) fetchSecurity(name string) (sec *portfoliov1.Security, err error) {
-	sec, err = svc.securities.Get(name)
-	if err != nil {
-		return nil, err
-	}
-
-	sec.ListedOn, err = svc.listedSecurities.List(name)
-	if err != nil {
-		return nil, err
-	}
-
-	return
 }
